API_files/bird-classifier/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
Pillow==8.2.0
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
